Ensure all cropped images have unique name
Previously I just pulled the filename described in the appropriate annotations file, however a small number of these were wrong and simply contained "%s" as the file name so some files may have been overwritten - evidently something went wrong when the annotation files were created. To counter this I now enumerate as I go through the list of files and prepend the counter to the filename.